feat(download): stable 512 buffer combo, for download

+++ app/scene/download/download.h
@@ -15,6 +15,6 @@ typedef struct {
 void scene_on_enter_download_progress(void* context);
 void scene_on_exit_download_progress(void* context);
 bool scene_on_event_download_progress(void* context, SceneManagerEvent event);
-void update_download_progress(App* app, int progress);
+void update_download_progress(App* app, size_t progress);
 
 #endif // DOWNLOAD_H
